refactor: 集中管理测试使用的mock数据
