+++ types/finance.d.ts
@@ -1,0 +1,8 @@
+interface Rule {
+  name: string
+  category: Category[]
+}
+interface Category {
+  name: string
+  source: string
+}
